# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kumande API Docs (Laravel)",
        "description": "API Documentation for Kumande Mobile & Web Apps",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/budget/by/{year}": {
            "get": {
                "tags": [
                    "Budget"
                ],
                "summary": "Get all budget plan in whole year",
                "operationId": "c98c97d5f4dc3ea27889df8adca21bfe",
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "description": "Budget year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2024"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Budget found"
                    },
                    "404": {
                        "description": "Budget not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Budget not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/budget/dashboard": {
            "post": {
                "tags": [
                    "Budget"
                ],
                "summary": "Get budget dashboard / summary",
                "operationId": "4b9e2ced6915be361533abdf6f47f3cc",
                "responses": {
                    "200": {
                        "description": "Budget found"
                    },
                    "404": {
                        "description": "Budget not found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/analytic/allergic": {
            "get": {
                "tags": [
                    "Analytic"
                ],
                "summary": "Get my allergic",
                "operationId": "fa0a2e7a2731f4194dbc45448c1cd0f5",
                "responses": {
                    "200": {
                        "description": "Allergic found"
                    },
                    "404": {
                        "description": "Allergic not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Allergic not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}